Ai improved, bugs fixed
Ai is now more reactionary, and is all around smarter. Need to work on
getting the health of the current pokemon it keeps changing. Once That
is finished all that is left is to figure out how to loop, figure out
how to deal with two pokemon dying in the same turn, and figure out how
to use outrage.